pyproject.toml
[project]
name = "energy_consumption"
version = "0.1.0"
description = "Time Series Forecasting with XGBoost, Hydra, DVC and MLflow"
authors = [
    {name = "Thiago Ribeiro", email = "152446274+thiagocarvalhor@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "pandas",
    "numpy",
    "scikit-learn",
    "xgboost",
    "matplotlib",
    "seaborn",
    "hydra-core",
    "mlflow",
    "rich",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pre-commit",
    "black",
    "isort",
    "flake8",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
